LayoutTests/http/tests/xmlhttprequest/simple-cross-origin-progress-events-expected.txt

Test that upload progress events are not dispatched for simple cross-origin requests (i.e. if the listener is set after calling send(), and there are no other reasons to make a preflight request).

Test 1: The URL is allowed for cross-origin requests
onprogress
onload
Response length: 8388608

Test 2: The URL is not allowed for cross-origin requests
onerror (expected)
Response length: 0
